[feature] Docker dev support
One can now use a complete docker environment to develop new feature
for this project, without having to care about installing local version
of node, meteor or mongodb. The project in mounted as a volume in the
container, and the docker-compose-dev.yml try to match the user and
group id to avoid permissions problems.

The meteor service also provide mongo-tools, which means that one can 
use and test the veritas-cli feature inside the container.