[fix] Don't both `fetchAsync()` and `observeChangesAsync()`
Closer reading of
https://docs.meteor.com/api/collections#Mongo-Cursor-observeAsync that
the latter does a superset of the former's job; we were therefore
racing with ourselves with unpredictable results (e.g. sometimes an
old subscription would send DDP updates from the past).

But before that, of course, I wasted a lot of time with the debugger
figuring out same from the source code :-P