[test][refactor]&[fix] a race condition when creating roles
Both the Meteor server and the various test preconditions used to call
`Roles.createRoleAsync`. It doesn't really matter that the latter did
proper `await`ing, because the check for the already-done condition
was imprecise (based on a count), with which the former would race.

- Given that `Roles.createRoleAsync` is a `Meteor.startup()`-time
task (also in production; and not reserved for testing) it is not a
fixture. Rename `loadFixtures` to `setupRoles`, and
`app/server/fixtures.js` to `app/server/roles.js`
- Since we all live in one and the same process, have all calls to
`setupRoles` consume one and the same promise
- Control for `duplicate key` separately for each role, for maximum
robustness
- Don't let `ResetDatabase` mess with roles anymore