[rewrite] Category management, part 1
“Categories” are slated to become a frontend-only concept. The
Kubernetes object should contain all the facts that are necessary and
sufficient to set up sites, so that the operator doesn't need to have
knowledge about the mapping between plugins and (EPFL-specific)
categories, plugins that are are always installed, and all that.

- For the time being, write both sets of data in created Kubernetes
objects. (As a reminder, we don't yet support updates.)
- Sprout `app/imports/api/k8s/siteCategories.js` with a bunch of OO
inside
- Sprout `(new Site(...)).getKubernetesPluginStruct()` method
- Rewrite `app/server/kubernetes.js` to go through that method